+++ l2m_core/__init__.py
@@ -1,0 +1,3 @@
+"""Lecture2Markdown Core Package."""
+
+__version__ = "1.1.2"